+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## OpaDotNet.Compilation v1.7.3 (2024-06-14)
+
+### Bug Fixes
+
+* Fix interop compiler to parse metadata when compiling from stream
+
 ## OpaDotNet.Compilation v1.7.2 (2024-06-13)
 
 ### Features